Move CoreLibLoader into stdlib_minimal
