+++ lib/src/metta/runner/stdlib_old.rs
@@ -2,8 +2,7 @@ use crate::*;
 use crate::space::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
-use crate::metta::text::SExprParser;
-use crate::metta::runner::{Metta, RunContext, ModuleLoader, ResourceKey};
+use crate::metta::runner::{Metta, RunContext, ResourceKey};
 use crate::metta::runner::string::Str;
 use crate::metta::types::{get_atom_types, get_meta_type};
 use crate::common::shared::Shared;
@@ -23,6 +22,7 @@ use rand::Rng;
 
 use super::arithmetics::*;
 use super::string::*;
+use super::stdlib_minimal::*;
 
 macro_rules! grounded_op {
     ($name:ident, $disp:literal) => {
@@ -1383,51 +1383,6 @@ impl CustomExecute for RandomFloatOp {
         let mut rng = rand::thread_rng();
         Ok(vec![Atom::gnd(Number::Float(rng.gen_range(range)))])
     }
-}
-
-use super::stdlib_minimal::*;
-use crate::metta::runner::METTA_CODE;
-
-/// Loader to Initialize the corelib module
-///
-/// NOTE: the corelib will be loaded automatically if the runner is initialized with one of the high-level
-/// init functions such as [Metta::new] and [Metta::new_with_stdlib_loader]
-#[derive(Debug)]
-pub(crate) struct CoreLibLoader;
-
-impl Default for CoreLibLoader {
-    fn default() -> Self {
-        CoreLibLoader
-    }
-}
-
-impl ModuleLoader for CoreLibLoader {
-    fn load(&self, context: &mut RunContext) -> Result<(), String> {
-        let space = DynSpace::new(GroundingSpace::new());
-        context.init_self_module(space, None);
-
-        register_rust_stdlib_tokens(&mut *context.module().tokenizer().borrow_mut());
-
-        let parser = SExprParser::new(METTA_CODE);
-        context.push_parser(Box::new(parser));
-
-        Ok(())
-    }
-}
-
-
-#[test]
-fn mod_space_op() {
-    let program = r#"
-        !(bind! &new_space (new-space))
-        !(add-atom &new_space (mod-space! stdlib))
-        !(get-atoms &new_space)
-    "#;
-    let runner = Metta::new(Some(runner::environment::EnvBuilder::test_env()));
-    let result = runner.run(SExprParser::new(program)).unwrap();
-
-    let stdlib_space = runner.module_space(runner.get_module_by_name("stdlib").unwrap());
-    assert_eq!(result[2], vec![Atom::gnd(stdlib_space)]);
 }
 
 #[cfg(test)]
